Move logic from config to Pali_Bot
